move usePlaylist() and useShareID()
