--- projects/frontend/src/graphql/client/types.local.ts
@@ -1,18 +1,0 @@
-export interface ILocalUserVariables {
-	userID: string
-}
-
-export interface ILocalUserData {
-	userID: string
-}
-
-export interface ILocalShareData {
-	shareID: string
-}
-
-export interface ILocalShareVariables {
-	shareID: string
-}
-export interface ITodoVariables {
-	id: string
-}
