--- projects/frontend/src/graphql/mutations/issue-auth-token.ts
@@ -1,15 +1,0 @@
-import gql from "graphql-tag"
-
-export interface IIssueAuthTokenVariables {
-	refreshToken: string
-}
-
-export interface IIssueAuthTokenData {
-	issueAuthToken: string
-}
-
-export const ISSUE_AUTH_TOKEN = gql`
-	mutation issueAuthToken($refreshToken: String!) {
-		issueAuthToken(refreshToken: $refreshToken)
-	}
-`
