+++ projects/frontend/src/hooks/data/useUpdateShareID.ts
@@ -1,0 +1,8 @@
+import { useIDStore } from "../../stores/idStore"
+import { useCallback } from "react"
+
+export const useUpdateShareID = () => {
+	const updateShareID = useIDStore(useCallback((state) => state.setShareID, []))
+
+	return updateShareID
+}
